Fix signed/unsigned comparison warnings

+++ fcntl.cc
@@ -273,7 +273,7 @@ TEST(Fcntl, SubRightNormalFD) {
   EXPECT_RIGHTS_EQ(&all, &rights);
   uint32_t fcntls;
   EXPECT_OK(cap_fcntls_get(fd, &fcntls));
-  EXPECT_EQ(CAP_FCNTL_GETFL, fcntls);
+  EXPECT_EQ((uint32_t)CAP_FCNTL_GETFL, fcntls);
 
   // Can't widen the subrights.
   EXPECT_NOTCAPABLE(cap_fcntls_limit(fd, CAP_FCNTL_GETFL|CAP_FCNTL_SETFL));
@@ -296,19 +296,19 @@ TEST(Fcntl, PreserveSubRights) {
   EXPECT_OK(cap_rights_get(fd, &cur_rights));
   EXPECT_RIGHTS_EQ(&rights, &cur_rights);
   EXPECT_OK(cap_fcntls_get(fd, &fcntls));
-  EXPECT_EQ(CAP_FCNTL_GETFL, fcntls);
+  EXPECT_EQ((uint32_t)CAP_FCNTL_GETFL, fcntls);
 
   // Limiting the top-level rights leaves the subrights unaffected...
   cap_rights_clear(&rights, CAP_READ);
   EXPECT_OK(cap_rights_limit(fd, &rights));
   EXPECT_OK(cap_fcntls_get(fd, &fcntls));
-  EXPECT_EQ(CAP_FCNTL_GETFL, fcntls);
+  EXPECT_EQ((uint32_t)CAP_FCNTL_GETFL, fcntls);
 
   // ... until we remove CAP_FCNTL.
   cap_rights_clear(&rights, CAP_FCNTL);
   EXPECT_OK(cap_rights_limit(fd, &rights));
   EXPECT_OK(cap_fcntls_get(fd, &fcntls));
-  EXPECT_EQ(0, fcntls);
+  EXPECT_EQ((uint32_t)0, fcntls);
   EXPECT_NOTCAPABLE(cap_fcntls_limit(fd, CAP_FCNTL_GETFL));
 
   close(fd);
@@ -388,9 +388,9 @@ TEST(Fcntl, OWNSubRights) {
   // Also check we can retrieve the subrights.
   uint32_t fcntls;
   EXPECT_OK(cap_fcntls_get(sock_get, &fcntls));
-  EXPECT_EQ(CAP_FCNTL_GETOWN, fcntls);
+  EXPECT_EQ((uint32_t)CAP_FCNTL_GETOWN, fcntls);
   EXPECT_OK(cap_fcntls_get(sock_set, &fcntls));
-  EXPECT_EQ(CAP_FCNTL_SETOWN, fcntls);
+  EXPECT_EQ((uint32_t)CAP_FCNTL_SETOWN, fcntls);
   // And that we can't widen the subrights.
   EXPECT_NOTCAPABLE(cap_fcntls_limit(sock_get, CAP_FCNTL_GETOWN|CAP_FCNTL_SETOWN));
   EXPECT_NOTCAPABLE(cap_fcntls_limit(sock_set, CAP_FCNTL_GETOWN|CAP_FCNTL_SETOWN));
